plugins: move code out of main.go into separate file
Plugin logic shouldn't clutter up main.go so much. Clean up by moving
it to plugins.go.
